+++ lib/src/html_data.dart
@@ -6,7 +6,6 @@ String htmlData({
   String? action,
   String? cData,
   required TurnstileOptions options,
-  required String onTurnstileReady,
   required String onTokenRecived,
   required String onTurnstileError,
   required String onTokenExpired,
@@ -36,8 +35,6 @@ String htmlData({
         return options.refreshExpired.name;
       case 'REFRESH_TIMEOUT':
         return options.refreshTimeout.name;
-      case 'READY':
-        return onTurnstileReady;
       case 'TOKEN_RECIVED':
         return onTokenRecived;
       case 'ERROR':
@@ -60,6 +57,7 @@ String _source = """
 
 <head>
    <meta charset="UTF-8">
+   <link rel="icon" href="data:,">
    <meta name="viewport"
       content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
    <script src="https://challenges.cloudflare.com/turnstile/v0/api.js?render=explicit"></script>
@@ -71,32 +69,42 @@ String _source = """
    <div id="cf-turnstile"></div>
    <script>
       turnstile.ready(function () {
-         <TURNSTILE_READY>
+           const widgetId = turnstile.render('#cf-turnstile', {
+              sitekey: '<TURNSTILE_SITE_KEY>',
+              action: '<TURNSTILE_ACTION>',
+              cData: '<TURNSTILE_CDATA>',
+              theme: '<TURNSTILE_THEME>',
+              size: '<TURNSTILE_SIZE>',
+              language: '<TURNSTILE_LANGUAGE>',
+              retry: '<TURNSTILE_RETRY>',
+              'retry-interval': parseInt('<TURNSTILE_RETRY_INTERVAL>'),
+              'refresh-expired': '<TURNSTILE_REFRESH_EXPIRED>',
+              'refresh-timeout': '<TURNSTILE_REFRESH_TIMEOUT>',
+              callback: function (token) {
+                 <TURNSTILE_TOKEN_RECIVED>
+              },
+              'error-callback': function (code) {
+                 <TURNSTILE_ERROR>
+              },
+              'expired-callback': function () {
+                 <TURNSTILE_TOKEN_EXPIRED>
+              }
+           });
 
-         const widgetId = turnstile.render('#cf-turnstile', {
-            sitekey: '<TURNSTILE_SITE_KEY>',
-            action: '<TURNSTILE_ACTION>',
-            cData: '<TURNSTILE_CDATA>',
-            theme: '<TURNSTILE_THEME>',
-            size: '<TURNSTILE_SIZE>',
-            language: '<TURNSTILE_LANGUAGE>',
-            retry: '<TURNSTILE_RETRY>',
-            'retry-interval': parseInt('<TURNSTILE_RETRY_INTERVAL>'),
-            'refresh-expired': '<TURNSTILE_REFRESH_EXPIRED>',
-            'refresh-timeout': '<TURNSTILE_REFRESH_TIMEOUT>',
-            callback: function (token) {
-               <TURNSTILE_TOKEN_RECIVED>
-            },
-            'error-callback': function (code) {
-               <TURNSTILE_ERROR>
-            },
-            'expired-callback': function () {
-               <TURNSTILE_TOKEN_EXPIRED>
-            }
-         });
-         
-         <TURNSTILE_CREATED>
-      });
+           <TURNSTILE_CREATED>
+        });
+
+      function getWidgetDimensions() {
+        const widgetElement = document.getElementById('cf-turnstile');
+        const rect = widgetElement.getBoundingClientRect();
+        const dimensions = {
+          width: rect.width,
+          height: rect.height
+        };
+
+        return JSON.stringify(dimensions) ;
+      };
+
    </script>
    <style>
       * {
@@ -108,4 +116,5 @@ String _source = """
 </body>
 
 </html>
+
 """;